Deleted VehicleModel.
Copied its content to ProductModel
